[ :wrench: |  | fix ] OrderQueryRepositoryTest 에러 수정
- payment 연관관계 관련 에러 수정

+++ src/test/resources/schema-h2.sql
@@ -24,7 +24,7 @@ create table cart
     member_id  bigint      null,
     product_id bigint      null,
     is_subs    bit         null,
-    quantity   int         null,
+    quantity   bigint      null,
     created_at datetime(6) null,
     updated_at datetime(6) null,
     deleted_at datetime(6) null
